Revert "iphone 8 fix attempt 1"
