+++ src/components/app/index.tsx
@@ -24,8 +24,181 @@ import styled from 'styled-components'
 import {useNotifications} from '../tools/use-notifications'
 import {HostApp} from './host'
 import {CreateEventApp} from './create-event'
-import {AppNavigation} from '../navigation'
+import {HOST_DOMAIN, PREFIX} from '../../config/network'
 
+const NavLink: React.FC<LinkProps> = (props) => {
+    const theme = useTheme()
+    return <Box clone color={theme.typography.body2.color}>
+        <Link {...props} style={{textDecoration: 'none'}}/>
+    </Box>
+}
+
+const drawerWidth = 200
+
+const useStyles = makeStyles((theme: Theme) =>
+    createStyles({
+        root: {
+            display: 'flex'
+        },
+        drawer: {
+            [theme.breakpoints.up('sm')]: {
+                width: drawerWidth,
+                flexShrink: 0
+            }
+        },
+        appBar: {
+            [theme.breakpoints.up('sm')]: {
+                width: `calc(100% - ${drawerWidth}px)`,
+                marginLeft: drawerWidth
+            }
+        },
+        menuButton: {
+            marginRight: theme.spacing(2),
+            [theme.breakpoints.up('sm')]: {
+                display: 'none'
+            }
+        },
+        // necessary for content to be below app bar
+        toolbar: theme.mixins.toolbar,
+        drawerPaper: {
+            paddingTop: 64,
+            boxSizing: 'border-box',
+            width: drawerWidth,
+            background: 'transparent',
+            border: 'none'
+        },
+        drawerPaper2: {
+            width: drawerWidth,
+            padding: theme.spacing(3)
+        },
+        content: {
+            flexGrow: 1,
+            padding: theme.spacing(3)
+        }
+    })
+)
+
+const AppNav: React.FC<GridProps> = ({children}) => {
+    const {cEvent, cUser} = useAppState()
+    const classes = useStyles()
+
+    const container = document.getElementById('root')
+
+    const [mobileOpen, setMobileOpen] = React.useState(false)
+
+    const handleDrawerToggle = () => {
+        setMobileOpen(!mobileOpen)
+    }
+    let drawer
+
+    if (cEvent.id === '-1' || cEvent.isNullEvent) {
+        drawer = null
+    } else {
+        drawer = (
+            <Box display='flex' flexDirection='column'>
+                {!cEvent.isNullEvent && cEvent.id !== '-1' &&
+                <NavLink to={`/event/${cEvent.id}`}
+                         onClick={() => setMobileOpen(false)}>
+                  <AdditionalText
+                    align='right'>
+                      {cEvent.name || 'К мероприятию'}
+                  </AdditionalText>
+                </NavLink>
+                }
+                {
+                    cEvent.isParticipating && !cUser.isNotAuthorized && cEvent.id !== '-1'
+                    && !cEvent.isFinished && cEvent.founderId !== cUser.id
+                    && !cEvent.notFound && <Fragment>
+                      <Box paddingTop={2}/>
+                      <NavLink to={`/team`} onClick={() => setMobileOpen(false)}>
+                        <AdditionalText align='right'>
+                          К команде
+                        </AdditionalText>
+                      </NavLink>
+                      <Box paddingTop={2}/>
+                      <NavLink to={`/feed`}
+                                onClick={() => setMobileOpen(false)}>
+                      <AdditionalText align='right'>
+                        К поиску
+                      </AdditionalText>
+                    </NavLink>
+                    </Fragment>
+                }
+                <Box paddingTop={2}/>
+                {cUser.id !== '-1' && !cUser.isNullUser &&
+                <NavLink to={`/user`} onClick={() => setMobileOpen(false)}>
+                  <AdditionalText align='right'>
+                    К себе
+                  </AdditionalText>
+                </NavLink>
+                }
+                <Box height='100px'/>
+                {/*<NavLink to={`/host`} onClick={() => setMobileOpen(false)}>*/}
+                {/*    <AdditionalText align='right'>*/}
+                {/*        Организаторам*/}
+                {/*    </AdditionalText>*/}
+                {/*</NavLink>*/}
+                {
+                    !cUser.isNotAuthorized && <NavLink to={`/event/create`}
+                                                       onClick={() => setMobileOpen(false)}>
+                      <AdditionalText align='right'>
+                        Создать мероприятие
+                      </AdditionalText>
+                    </NavLink>
+                }
+            </Box>
+        )
+    }
+    return <div className={classes.root}>
+        <Hidden smUp implementation="css">
+            <AppBar position="fixed" className={classes.appBar}>
+                <Toolbar>
+                    <IconButton
+                        color="inherit"
+                        aria-label="open drawer"
+                        edge="start"
+                        onClick={handleDrawerToggle}
+                        className={classes.menuButton}
+                    >
+                        Меню
+                    </IconButton>
+                </Toolbar>
+            </AppBar>
+        </Hidden>
+        <nav className={classes.drawer} aria-label="mailbox folders">
+            {/* The implementation can be swapped with js to avoid SEO duplication of links. */}
+            <Hidden smUp implementation="css">
+                <Drawer className={classes.drawer}
+                        container={container}
+                        variant="temporary"
+                        anchor={'left'}
+                        open={mobileOpen}
+                        onClose={handleDrawerToggle}
+                        classes={{
+                            paper: classes.drawerPaper2
+                        }}
+                        ModalProps={{
+                            keepMounted: true // Better open performance on mobile.
+                        }}
+                >
+                    {drawer}
+                </Drawer>
+            </Hidden>
+            <Hidden xsDown implementation="css">
+                <Drawer
+                    classes={{
+                        paper: classes.drawerPaper
+                    }}
+                    variant="permanent"
+                    open
+                >
+                    {drawer}
+                </Drawer>
+            </Hidden>
+        </nav>
+        {children}
+    </div>
+}
 
 const StyledContainer = styled(Container)`
   min-height: 100vh;
@@ -62,53 +235,53 @@ export const App: React.FC = () => {
 
     return <Switch>
         <Route path='/event/create'>
-            <AppNavigation>
+            <AppNav>
                 <RootContainer>
                     <CreateEventApp/>
                 </RootContainer>
-            </AppNavigation>
+            </AppNav>
         </Route>
         <Route path='/user/:userId'>
-            <AppNavigation>
+            <AppNav>
                 <RootContainer>
                     <UserApp/>
                 </RootContainer>
-            </AppNavigation>
+            </AppNav>
         </Route>
         <Route path='/user'>
-            <AppNavigation>
+            <AppNav>
                 <RootContainer>
                     <UserApp/>
                 </RootContainer>
-            </AppNavigation>
+            </AppNav>
         </Route>
         <Route path='/event/:eventId'>
-            <AppNavigation>
+            <AppNav>
                 <RootContainer>
                     <EventApp/>
                 </RootContainer>
-            </AppNavigation>
+            </AppNav>
         </Route>
         <Route path='/event'>
-            <AppNavigation>
+            <AppNav>
                 <RootContainer>
                     <EventApp/>
                 </RootContainer>
-            </AppNavigation>
+            </AppNav>
         </Route>
         <Route path='/team'>
-            <AppNavigation>
+            <AppNav>
                 <RootContainer>
                     <TeamApp/>
                 </RootContainer>
-            </AppNavigation>
+            </AppNav>
         </Route>
         <Route path='/feed'>
-            <AppNavigation>
+            <AppNav>
                 <RootContainer>
                         <FeedApp/>
                 </RootContainer>
-            </AppNavigation>
+            </AppNav>
         </Route>
         <Route path='/host'>
             <HostApp/>
