--- src/components/navigation/loading.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import {Grid} from '@material-ui/core'
-
-export const LoadingMenu: React.FC = () => {
-    return <Grid container>
-    </Grid>
-}
